+++ car_rent.Server/DataProvider/CarRentalDataProvider2.cs
@@ -16,34 +16,17 @@
     using System.Net;
     using System.Text.Json.Serialization;
 
-    public class RentInfoFromApi
-    {
-        [JsonPropertyName("rentId")]
-        public string RentId { get; set; }
-        [JsonPropertyName("start")]
-        public DateTime StartDate { get; set; }
-        [JsonPropertyName("end")]
-        public DateTime EndDate { get; set; }
-
-        [JsonPropertyName("carBrand")]
-        public string CarBrand { get; set; }
-        [JsonPropertyName("carModel")]
-        public string CarModel { get; set; }
-        [JsonPropertyName("carYear")]
-        public int CarYear { get; set; }
-        [JsonPropertyName("price")]
-        public float Price { get; set; }
-    }
-
+   
     public class CarRentalDataProvider2 : ICarRentalDataProvider
     {
         private readonly HttpClient _httpClient;
         private readonly string _accessToken;
-        private readonly string _apiUrl = "http://localhost:5173";
+        private readonly string _apiUrl2;
 
-        public CarRentalDataProvider2(HttpClient httpClient)
+        public CarRentalDataProvider2(HttpClient httpClient, string apiUrl2)
         {
             _httpClient = httpClient;
+            _apiUrl2 = apiUrl2;
         }
 
         public string GetProviderName()
@@ -71,7 +54,7 @@
             if (!result) return null;
 
             // Proceed with the rent car operation
-            var rentCarResponse = await _httpClient.GetAsync($"{_apiUrl}/api/Offer/rentCar/{offerId}/{clientId}");
+            var rentCarResponse = await _httpClient.GetAsync($"{_apiUrl2}/api/Offer/rentCar/{offerId}/{clientId}");
             if (!rentCarResponse.IsSuccessStatusCode)
             {
                 return null;
@@ -84,13 +67,13 @@
 
         public async Task<bool> ReturnCar(string rentId)
         {
-            var rentCarResponse = await _httpClient.GetAsync($"{_apiUrl}/api/Rent/readyToReturn/{rentId}");
+            var rentCarResponse = await _httpClient.GetAsync($"{_apiUrl2}/api/Rent/readyToReturn/{rentId}");
             return rentCarResponse.IsSuccessStatusCode;
         }
 
         public async Task<OfferFromApi?> GetOneOfferFromApi(string offerId)
         {
-            var offerResponse = await _httpClient.GetAsync($"{_apiUrl}/api/Offer/id/{offerId}");
+            var offerResponse = await _httpClient.GetAsync($"{_apiUrl2}/api/Offer/id/{offerId}");
             if (!offerResponse.IsSuccessStatusCode)
             {
                 return null;
@@ -103,7 +86,7 @@
 
         public async Task<List<OfferFromApi>> GetOfferToDisplays(DateTime startDate, DateTime endDate, string search_brand, string search_model, string clientId, string email)
         {
-            var requestUrl = $"{_apiUrl}/api/offer?startDate={startDate:yyyy-MM-dd}&endDate={endDate:yyyy-MM-dd}&brand={search_brand}&model={search_model}&clientId={clientId}";
+            var requestUrl = $"{_apiUrl2}/api/offer?startDate={startDate:yyyy-MM-dd}&endDate={endDate:yyyy-MM-dd}&brand={search_brand}&model={search_model}&clientId={clientId}";
 
             try
             {
@@ -126,7 +109,7 @@
         private async Task<bool> CheckIfClientExistsInApi(string clientId, ApplicationUser user)
         {
             // Check if the user exists in the external API
-            var checkClientResponse = await _httpClient.GetAsync($"{_apiUrl}/api/Client/checkClient/{clientId}");
+            var checkClientResponse = await _httpClient.GetAsync($"{_apiUrl2}/api/Client/checkClient/{clientId}");
             if (checkClientResponse.StatusCode == HttpStatusCode.NotFound)
             {
                 // Prepare and send user information to the external API
@@ -136,7 +119,7 @@
                 var userInformationJson = JsonSerializer.Serialize(userInformation);
                 var content = new StringContent(userInformationJson, Encoding.UTF8, "application/json");
 
-                var createClientResponse = await _httpClient.PostAsync($"{_apiUrl}/api/Client/createClient", content);
+                var createClientResponse = await _httpClient.PostAsync($"{_apiUrl2}/api/Client/createClient", content);
                 if (!createClientResponse.IsSuccessStatusCode)
                 {
                     return false;
